chore: refactor lock file
